Extract accounting reconciliation view model builders
